Backend/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-cors
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
wtforms
email-validator
flask-jwt-extended
flask-restx
requests
flask[async]
beautifulsoup4
pandas